Further relocation from 'native' to 'kotlin' Moved the left behind files and renamed the package.
