+++ backend.native/compiler/ir/backend.native/src/org/jetbrains/kotlin/backend/konan/DescriptorUtils.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.kotlin.backend.native
+package org.jetbrains.kotlin.backend.konan
 
 import org.jetbrains.kotlin.descriptors.ClassDescriptor
 import org.jetbrains.kotlin.descriptors.ClassKind
