+++ backend.native/compiler/ir/backend.native/src/org/jetbrains/kotlin/backend/konan/NativeIrCodegenFactory.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.kotlin.backend.native
+package org.jetbrains.kotlin.backend.konan
 
 import org.jetbrains.kotlin.codegen.*
 import org.jetbrains.kotlin.codegen.context.PackageContext
